backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2

python-jose[cryptography]
pydantic
requests